Task: 'About' links to an anchor tag instead of new page
